requirements.txt
#
# This file is autogenerated by pip-compile with python 3.11
# To update, run:
#
#    pip-compile
#
appnope==0.1.3
    # via ipython
arabic-reshaper==2.1.4
    # via xhtml2pdf
asgiref==3.5.2
    # via django
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   pyhanko
    #   pyhanko-certvalidator
asttokens==2.1.0
    # via stack-data
async-timeout==4.0.2
    # via redis
attrs==22.1.0
    # via pytest
backcall==0.2.0
    # via ipython
certifi==2022.9.24
    # via
    #   -r requirements.in
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via pyhanko
coverage==6.5.0
    # via -r requirements.in
crispy-bootstrap5==0.7
    # via -r requirements.in
cryptography==38.0.3
    # via
    #   pyhanko
    #   pyhanko-certvalidator
cssselect2==0.7.0
    # via svglib
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via odfpy
deprecated==1.2.13
    # via redis
diff-match-patch==20200713
    # via django-import-export
django==4.1.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-admin-autocomplete-filter
    #   django-browser-reload
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-import-export
    #   django-redis
    #   django-user-visit
    #   djangorestframework
django-admin-autocomplete-filter==0.7.1
    # via -r requirements.in
django-admin-rangefilter==0.9.0
    # via -r requirements.in
django-browser-reload==1.6.0
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via crispy-bootstrap5
django-debug-toolbar==3.7.0
    # via -r requirements.in
django-environ==0.9.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-import-export==3.0.1
    # via -r requirements.in
django-redis==5.2.0
    # via -r requirements.in
django-simple-history==3.2.0
    # via -r requirements.in
django-user-visit==1.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
executing==1.2.0
    # via stack-data
future==0.18.2
    # via arabic-reshaper
hiredis==2.0.0
    # via -r requirements.in
html5lib==1.1
    # via xhtml2pdf
idna==3.4
    # via
    #   -r requirements.in
    #   requests
iniconfig==1.1.1
    # via pytest
ipython==8.6.0
    # via -r requirements.in
jedi==0.18.1
    # via ipython
lxml==4.9.1
    # via svglib
markuppy==1.14
    # via tablib
matplotlib-inline==0.1.6
    # via ipython
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
oscrypto==1.3.0
    # via pyhanko-certvalidator
packaging==21.3
    # via
    #   pytest
    #   pytest-sugar
    #   redis
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.3.0
    # via
    #   reportlab
    #   xhtml2pdf
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.32
    # via ipython
psycopg2==2.9.5
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via ipython
pyhanko==0.15.1
    # via xhtml2pdf
pyhanko-certvalidator==0.19.6
    # via
    #   pyhanko
    #   xhtml2pdf
pyparsing==3.0.9
    # via packaging
pypdf3==1.0.6
    # via xhtml2pdf
pytest==7.2.0
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-sugar
pytest-django==4.5.2
    # via -r requirements.in
pytest-sugar==0.9.6
    # via -r requirements.in
python-bidi==0.4.2
    # via xhtml2pdf
pytz==2022.6
    # via
    #   djangorestframework
    #   pyhanko
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   pyhanko
    #   tablib
qrcode==7.3.1
    # via pyhanko
redis==4.3.4
    # via
    #   -r requirements.in
    #   django-redis
reportlab==3.6.12
    # via
    #   svglib
    #   xhtml2pdf
requests==2.28.1
    # via
    #   pyhanko
    #   pyhanko-certvalidator
six==1.16.0
    # via
    #   asttokens
    #   html5lib
    #   python-bidi
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.0
    # via ipython
svglib==1.4.1
    # via xhtml2pdf
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
termcolor==2.1.0
    # via
    #   -r requirements.in
    #   pytest-sugar
tinycss2==1.2.1
    # via
    #   cssselect2
    #   svglib
tqdm==4.64.1
    # via pypdf3
traitlets==5.5.0
    # via
    #   -r requirements.in
    #   ipython
    #   matplotlib-inline
tzdata==2022.6
    # via pytz-deprecation-shim
tzlocal==4.2
    # via pyhanko
ua-parser==0.16.1
    # via user-agents
uritools==4.0.0
    # via pyhanko-certvalidator
urllib3==1.26.12
    # via requests
user-agents==2.2.0
    # via django-user-visit
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
whitenoise==6.2.0
    # via -r requirements.in
wrapt==1.14.1
    # via deprecated
xhtml2pdf==0.2.8
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
